+++ src/client/index.iam.ts
@@ -11,7 +11,7 @@ import {
 import { signRequest } from './aws-sigv4.js';
 
 export { Database } from './Database.js';
-export { Collection } from './Collection.js';
+export { Collection, type InputBase, type ResultBase } from './Collection.js';
 export { FindCursor } from './FindCursor.js';
 
 /**